Web module is added. Fixed root module.
