--- src/main/java/sq/vk/configuration/mvc/DispatcherServletInitializer.java
@@ -1,33 +1,0 @@
-package sq.vk.configuration.mvc;
-
-import javax.servlet.ServletRegistration;
-
-import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
-
-/**
- * Created by Vadzim Kavalkou on 23.03.2016.
- *
- * Dispatcher servlet configuration.
- */
-public class DispatcherServletInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
-
-    @Override
-    protected Class<?>[] getRootConfigClasses() {
-        return new Class<?>[]{RootConfig.class};
-    }
-
-    @Override
-    protected Class<?>[] getServletConfigClasses() {
-        return new Class<?>[]{ServletConfig.class};
-    }
-
-    @Override
-    protected String[] getServletMappings() {
-        return new String[]{"/"};
-    }
-
-    @Override
-    public void customizeRegistration(ServletRegistration.Dynamic registration) {
-        registration.setInitParameter("throwExceptionIfNoHandlerFound", "true");
-    }
-}
