--- src/main/java/sq/vk/configuration/security/SecurityWebApplicationInitializer.java
@@ -1,12 +1,0 @@
-package sq.vk.configuration.security;
-
-import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
-
-/**
- * Created by Vadzim Kavalkou on 13.11.2016.
- */
-public class SecurityWebApplicationInitializer extends AbstractSecurityWebApplicationInitializer {
-
-    public SecurityWebApplicationInitializer() {
-    }
-}
