+++ statusquo-web/src/main/java/sq/vk/web/configuration/mvc/RootConfig.java
@@ -1,0 +1,17 @@
+package sq.vk.web.configuration.mvc;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+/**
+ * Created by Vadzim Kavalkou on 23.03.2016.
+ *
+ * Root configuration.
+ */
+@Configuration
+@ComponentScan(basePackages = {"sq.vk"},
+        excludeFilters = {@ComponentScan.Filter(type = FilterType.ANNOTATION, value = EnableWebMvc.class)})
+public class RootConfig {
+}
