+++ statusquo-web/src/main/java/sq/vk/web/configuration/security/SecurityWebApplicationInitializer.java
@@ -1,0 +1,12 @@
+package sq.vk.web.configuration.security;
+
+import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
+
+/**
+ * Created by Vadzim Kavalkou on 13.11.2016.
+ */
+public class SecurityWebApplicationInitializer extends AbstractSecurityWebApplicationInitializer {
+
+    public SecurityWebApplicationInitializer() {
+    }
+}
