+++ statusquo-web/src/main/resources/properties/jdbc.properties
