+++ statusquo-web/src/main/resources/sql/create-clients.sql
