+++ statusquo-web/src/main/resources/sql/create-gameinfo.sql
