+++ statusquo-web/src/main/resources/sql/create-statistics-gameinfo.sql
