+++ statusquo-web/src/main/resources/sql/init-clients.sql
