+++ statusquo-web/src/main/resources/sql/init-gameinfo.sql
