+++ statusquo-web/src/main/resources/sql/init-statistics.sql
